eliminate "flicker" when channel list is open with non-standard height
we need to account both for scenario when no data is available when displaying
the view (i.e. channel list has not been received yet) and when we already have it.
In the latter case the list would briefly appear with standard height and then
reload to preset height. This commit eliminates that unwanted behaviour.